Reformat & Undraft the Uses page (#6)
Because "Uses" was a draft but still appeared on the main menu, it
was basically a dead link. I'll need to look into why the checker didn't
catch that.

Also does some small Markdown reformatting changes, to change the
header levels on the page and do some re-indenting.

This page is nowhere near done, I'm just unbreaking a link. I'll
actually finish this later.