Write post about design
I've had this in mind for a while, and it was good to finally get it
written down. This is kind of in the vein of things that I originally
planned on using a blog for (rather than egg sandwich review /
television shows), and I'm actually rather proud of it. I think it's
decently written, and it covers a (loosely) technology related topic
that I care about. So that's nice.