Add author's note on She-Ra post (#27)
* Add author's note on She-Ra post

Adds a note saying that I know the She-Ra post as it stands is pretty
bad.

This is a way, way larger PR than I wanted this to be. This was supposed
to be a very, very small pr of maybe just the note. But I had updated
the hugo binary in the meantime, which brought in some significant
breaking changes. The main compensation for those was updating my inline
HTML into stuff that wasn't that.

Also updated the theme, which *also* had some breaking changes
(specifically with regard to a new icon font).

* Also update hugo version in CircleCI